aggiunto CarType enum e aggiunti errori

+++ src/main/java/co/develhope/eserciziocrud/controller/CarController.java
@@ -1,11 +1,15 @@
 package co.develhope.eserciziocrud.controller;
 
 import co.develhope.eserciziocrud.entities.Car;
+import co.develhope.eserciziocrud.entities.Enum.CarType;
 import co.develhope.eserciziocrud.repositories.CarRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/cars")
@@ -25,24 +29,34 @@ public class CarController {
     }
 
     @GetMapping("/{id}")
-    public Car getCarById(@PathVariable Long id) {
-        return carRepository.findById(id)
-                .orElse(new Car());
+    public ResponseEntity<Car> getCarById(@PathVariable Long id) {
+        Optional<Car> car = carRepository.findById(id);
+        if (car.isPresent()) {
+            return ResponseEntity.ok(car.get());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new Car());
+        }
     }
 
     @PutMapping("/{id}")
-    public Car updateCarType(@PathVariable Long id, @RequestParam String type) {
-        Car car = carRepository.findById(id).orElse(null);
-        if (car == null) {
-            return null;
+    public ResponseEntity<Car> updateCarType(@PathVariable Long id, @RequestParam CarType type) {
+        if (carRepository.existsById(id)) {
+            Car car = carRepository.findById(id).get();
+            car.setType(type);
+            return ResponseEntity.ok(carRepository.save(car));
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new Car());
         }
-        car.setType(type);
-        return carRepository.save(car);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCarById(@PathVariable Long id) {
-        carRepository.deleteById(id);
+    public ResponseEntity<Void> deleteCar(@PathVariable Long id) {
+        if (carRepository.existsById(id)) {
+            carRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
     }
 
     @DeleteMapping
